Move bib cache to a new lib file
